Adding specs (#69)
This project starts from a blank app and builds up an industrial grade Photogram. These tests aim to match the functionality and look of the app at the end of the [final video for this project](https://canvas.uchicago.edu/courses/47618/pages/video-photogram-industrial-pt-5).

The [target for this project](https://pg-industrial-2-final.herokuapp.com/) has lot more going on with the interface. The specs are intended to pass for both the end of the video and the target app, without testing all the functionality of the target. Another PR is open for advanced backend specs that could also be added.

Specs here:

- [x] basic Devise sign in
- [x] `before_action :authenticate_user!` in ApplicationController
- [x] presence of bootstrap navbar
- [x] user account links in navbar
- [x] presence of `<div class="card">` for photo partial
- [x] ability to visit a user show page with `/USERNAME`
- [x] presence of comments within a photo card
- [x] ability to add a comment with the form